Added command rcmstatuslist
This is command had some interesting challenges.  I needed to get the RCM ID
from the libraryStatus command.  And the way to get that is to look for a
particular string in the name (RCM Spectra PC).  Then I had to strip that (along
with a colon) to get the FR<integer>/RCM portion of the name.  And then this was
passed as the argument to the RCMStatus action.