Added command gettapstate
Attempt to get the terapack access port (TAP) status of each TAPDevice and
drawer.  Here are the things that are odd to me:
1. For mainTop for drawers 1 -> 14 and mainBottom for drawers 1 -> 14, I only get
two elements: doorOpen and magazinePresent.  [magazineSeated, magazineType, and
rotaryPosition are missing]   Should I be getting the elements that are missing?
(Seems like I should be getting the magazineType??)
2. For leftBulk for drawers 1-> 14, rightBulk for drawers 1->14, I get exception
from the command.  Is this okay for my library configuration?

Sent note to Spectra about this.  Put <unknown> string for the items that aren't
returned