Added commands getcanlog and getcanlognames
getcanlog: Retrieves the specified zip file containing Controller Area
Network (CAN) logs from the Library Control Module (LCM).

getcanlognames: Returns a list of the zip files containing the Controller
Area Network (CAN) logs that are currently stored in the Library Control
Module (LCM). The CAN logs collected for each day are zipped and stored
on the hard  drive in the LCM. Each zip filename includes the date it
was created.