Added commands getkernellog and getkernellognames
getkernellog: Retrieves the specified zip file containing kernel logs
from the Library Control Module (LCM).

getkernellognames: Returns a list of the zip files containing the kernel
logs that are currently stored in the Library Control Module (LCM). The
kernel logs collected for each day are zipped and stored on the hard
drive in the LCM. Each zip filename includes the date it was created.