Added command: gettrace <trace type>
Retrieves the ASCII formatted data for the type of trace specified by the command.
Outputs a filename in the current working directory that contains the trace data in ASCII format.

Output filename format: command_YYYY-MM-DD_<unique number starting at 1>

Here are the types of traces:
choices=['Action', 'AutoDriveClean', 'AutoSupport', 'BackgroundClient',
         'CAN', 'Connection', 'Encryption', 'Error', 'EtherLib',
         'Event', 'Geometry', 'GPIO', 'HHM', 'HydraExit',
         'Initialization', 'Inventory', 'Kernel', 'Lock',
         'LogicalLibrary', 'Message', 'MLM', 'Motion',
         'MotionInventory', 'MotionOptions', 'MotionRestart1',
         'MotionRestart2', 'PackageUpdate', 'Pools', 'SNMP',
         'WebServer']

The documentation is very poor about defining all of the above.

I made it so that the user could type the lower case of the option.