Github Issue #4
The security audit stopped working with BlueScale 12.8.03.
We were looking for the string
"Security audit is not running."
and in the new version it was modified to
"Security Audit is not running."
We were also looking for that string in multiple places in the
code.  We refactored this and now doing a case-insensitive search.