Detect that Security Audit Completed
The newer BlueScale 12.8.03 prints out a new message when the security
audit is complete "Security audit completed..." Allow this message to
detect that the security audit is not running anymore.

We might consider changing the default polarity of our checks from
running to not running. Then we only have to look for the string
Security Audit is X percent complete.

For now just make this simple fix.