Add ability to move tapes to/from slot/ee/drive
BlueScale 12.8.04.01 added the ability to move tapes to/from
slot/ee/drive. slapi can not take advantage of this command. The initial
version tries to simplify the input by only allowing the source to be a
barcode. The new command is inventorymove.

In addition, we also print out the SourceSlot and SourceOffset in the
inventorylist command. This should help give the user the ability to see
which slot a tape came from when it is mounted in a drive.